FrontBaseController moved from Components to project-base Controllers
- FrontBaseController is just antoher Controller and as Component does not really make sense.
