+++ project-base/src/Shopsys/ShopBundle/Controller/Front/RobotsController.php
@@ -2,7 +2,6 @@
 
 namespace Shopsys\ShopBundle\Controller\Front;
 
-use Shopsys\FrameworkBundle\Component\Controller\FrontBaseController;
 use Shopsys\FrameworkBundle\Component\Domain\Domain;
 use Shopsys\FrameworkBundle\Component\Sitemap\SitemapService;
 use Symfony\Component\HttpFoundation\Response;
