+++ project-base/src/Shopsys/ShopBundle/Controller/Test/ErrorHandlerController.php
@@ -3,7 +3,7 @@
 namespace Shopsys\ShopBundle\Controller\Test;
 
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
-use Shopsys\FrameworkBundle\Component\Controller\FrontBaseController;
+use Shopsys\ShopBundle\Controller\Front\FrontBaseController;
 use Symfony\Component\HttpFoundation\Response;
 
 class ErrorHandlerController extends FrontBaseController
